+++ components/Modals/AddFolderModal/AddFolderModal.tsx
@@ -43,46 +43,6 @@ const AddFolderModal = ({ setOpenAddFolderModal }: IAddFolderModalProps) => {
         </button>
       </div>
     </ModalLayout>
-    // <ModalPortalWrapper id="add-folder-portal">
-    //   <div
-    //     className={styles.overlay}
-    //     onClick={() => setOpenAddFolderModal(false)}
-    //   >
-    //     <div
-    //       className={styles.modalWrapper}
-    //       onClick={(e) => e.stopPropagation()}
-    //     >
-    //       <div
-    //         className={styles.closeWrapper}
-    //         onClick={() => setOpenAddFolderModal(false)}
-    //       >
-    //         <Image
-    //           className={styles.image}
-    //           fill
-    //           src="/assets/modal-close.svg"
-    //           alt="close modal"
-    //         />
-    //       </div>
-    //       <div className={styles.contents}>
-    //         <h3 className={styles.title}>폴더 추가</h3>
-    //         <input
-    //           ref={folderNameRef}
-    //           className={styles.input}
-    //           type="text"
-    //           placeholder="내용 입력"
-    //         />
-    //         <div className={styles.buttonWrapper}>
-    //           <button
-    //             className={styles.addFolderButton}
-    //             onClick={handleClickAddFolder}
-    //           >
-    //             추가하기
-    //           </button>
-    //         </div>
-    //       </div>
-    //     </div>
-    //   </div>
-    // </ModalPortalWrapper>
   );
 };
