+++ components/Modals/AddLinkModal/AddLinkModal.tsx
@@ -3,11 +3,11 @@
 import { useState } from "react";
 
 import ModalLayout from "../ModalLayout";
-import styles from "./AddModal.module.scss";
+import styles from "./AddLinkModal.module.scss";
 import FolderItem from "./FolderItem";
 
 interface IAddModalProps {
-  setOpenAddModal: React.Dispatch<React.SetStateAction<boolean>>;
+  setOpenAddLinkModal: React.Dispatch<React.SetStateAction<boolean>>;
   selectedLinkValue: string;
 }
 
@@ -18,8 +18,11 @@ const MOCK_FOLDERS: { name: string; length: number }[] = [
   { name: "나만의 장소", length: 3 },
 ];
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const AddModal = ({ setOpenAddModal, selectedLinkValue }: IAddModalProps) => {
+const AddLinkModal = ({
+  setOpenAddLinkModal,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  selectedLinkValue,
+}: IAddModalProps) => {
   // TODO: api로 폴더 데이터를 받아오면 MOCK 데이터 삭제 예정
 
   const [checkedItemId, setCheckedItemId] = useState(-1);
@@ -28,14 +31,14 @@ const AddModal = ({ setOpenAddModal, selectedLinkValue }: IAddModalProps) => {
     // TODO: POST request : selectedLinkValue를 checkedItemId 를 가진 폴더에 저장
     // TODO: 바로 닫히지 않고, 데이터가 잘 저장되었다고 알려주는 텀이 있으면 좋겠다.
     setTimeout(() => {
-      setOpenAddModal(false);
+      setOpenAddLinkModal(false);
     }, 500);
   };
 
   return (
     <ModalLayout
       portalId="add-portal"
-      handleClickCloseModal={() => setOpenAddModal(false)}
+      handleClickCloseModal={() => setOpenAddLinkModal(false)}
     >
       <h3 className={styles.title}>폴더에 추가</h3>
       <h4 className={styles.subTitle}>링크 주소</h4>
@@ -57,4 +60,4 @@ const AddModal = ({ setOpenAddModal, selectedLinkValue }: IAddModalProps) => {
   );
 };
 
-export default AddModal;
+export default AddLinkModal;
